+++ TemplateMultiTenant.Auth/ViewModel/ClientModel.cs
@@ -1,6 +1,7 @@
 ﻿//Classe POCO contendo propriedades necessárias para autenticação do client
 using TemplateMultiTenant.Auth.Models;
 using System.ComponentModel.DataAnnotations;
+using System;
 
 namespace TemplateMultiTenant.Auth.ViewModel
 {
@@ -12,8 +13,8 @@ namespace TemplateMultiTenant.Auth.ViewModel
         [Required]
         public string Name { get; set; }
 
-        [Required]
-        public int CNPJ { get; set; }
+        [Required]                
+        public long CNPJ { get; set; }
 
         [Required]
         public SubscriptionType SubscriptionType { get; set; }
